#11 Use a hook to wait for results
