+++ src/components/Parallax.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import propTypes from 'prop-types';
 import styled from 'styled-components';
-import { graphql, useStaticQuery } from 'gatsby';
 
 const ParallaxDiv = styled.div`
     aspect-ratio: 2309/1536;
@@ -19,41 +18,10 @@ const ParallaxDiv = styled.div`
     }
 `;
 
-const getImageURLs = () => {
-  const data = useStaticQuery(
-    graphql`
-      query {
-        climateSimulator: file(relativePath: { eq: "climate_simulator.jpg" }) {
-          childImageSharp {
-            fluid(quality: 100) {
-              ...GatsbyImageSharpFluid_withWebp
-            }
-          }
-        }
-        starlightLounge: file(
-          relativePath: { eq: "starlight_lounge.jpg" }
-        ) {
-          childImageSharp {
-            fluid(quality: 100) {
-              ...GatsbyImageSharpFluid_withWebp
-            }
-          }
-        }
-      }
-    `,
-  );
-
-  return {
-    'Climate Simulator': data.climateSimulator.childImageSharp.fluid.srcWebp,
-    'Starlight Lounge': data.starlightLounge.childImageSharp.fluid.srcWebp,
-  };
-};
-
-export default function Parallax({ image }) {
-  const images = getImageURLs();
-  return <ParallaxDiv url={images[image]} />;
+export default function Parallax({ url }) {
+  return <ParallaxDiv url={url} />;
 }
 
 Parallax.propTypes = {
-  image: propTypes.oneOf(['Climate Simulator', 'Starlight Lounge']).isRequired,
+  url: propTypes.string.isRequired,
 };
